requirements.txt
flask
pyowm
python-dotenv